Reorganise the stuff in the form of a library
It took quite a while to learn about the nuances
of import system in python. I have decided that for
now a user can avail the functionality by simply
modifyig the sys.path .
In future I intend to restructure it just as
scikit-learn is structured (relative imports like
from ..utils import bench) or maybe something better?